infrastructure_api/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
